move MethodSupport and public access to method
