+++ jfr-daemon/src/main/java/com/newrelic/jfr/daemon/JFRUploader.java
@@ -7,12 +7,18 @@
 
 package com.newrelic.jfr.daemon;
 
+import com.newrelic.telemetry.events.Event;
 import com.newrelic.telemetry.events.EventBatch;
 import com.newrelic.telemetry.metrics.MetricBatch;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.time.Instant;
+import java.util.List;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
+
 import jdk.jfr.consumer.RecordingFile;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -88,10 +94,41 @@ public final class JFRUploader {
     EventBatch eventBatch = bufferedMetrics.createEventBatch();
     if (!eventBatch.isEmpty()) {
       logger.info(String.format("Sending events batch of size %s", eventBatch.size()));
+
+      List<Event> methodSampleEvents = eventBatch.getTelemetry().stream()
+              .filter(event -> event.getEventType().contains("MethodSample"))
+              .collect(Collectors.toList());
+
+      List<Event> flameLevelEvents = eventBatch.getTelemetry().stream()
+              .filter(event -> event.getEventType().contains("Flame"))
+              .collect(Collectors.toList());
+      
+      int methodSampleThreadRunCount =  eventBatch.getTelemetry().stream()
+              .filter(event -> event.getEventType().contains("MethodSample"))
+              .mapToInt(event -> countMatches(event.getAttributes().asMap().get("stackTrace").toString(), "java.lang.Thread.run()"))
+              .sum();
+
+      int flamelevelThreadCount=  eventBatch.getTelemetry().stream()
+              .filter(event -> event.getEventType().contains("Flame"))
+              .filter(event -> event.getAttributes().asMap().get("flamelevel.name").toString().contains("Thread.run"))
+              .mapToInt(event -> Integer.parseInt(event.getAttributes().asMap().get("flamelevel.value").toString())).sum();
       telemetrySender.sendBatch(eventBatch);
+
     }
   }
 
+  /* Counts how many times the substring appears in the larger string. */
+  public static int countMatches(String text, String str) {
+    Matcher matcher = Pattern.compile(str).matcher(text);
+
+    int count = 0;
+    while (matcher.find()) {
+      count++;
+    }
+
+    return count;
+  }
+  
   RecordingFile openRecordingFile(Path file) {
     try {
       return new RecordingFile(file);
