+++ jfr-mappers/src/main/java/com/newrelic/jfr/toevent/JITCompilationMapper.java
@@ -10,7 +10,7 @@ package com.newrelic.jfr.toevent;
 import static com.newrelic.jfr.RecordedObjectValidators.*;
 
 import com.newrelic.jfr.Workarounds;
-import com.newrelic.jfr.profiler.MethodSupport;
+import com.newrelic.jfr.MethodSupport;
 import com.newrelic.telemetry.Attributes;
 import com.newrelic.telemetry.events.Event;
 import java.time.Duration;
