+++ jfr-mappers/src/main/java/com/newrelic/jfr/toevent/ValhallaVBCDetector.java
@@ -2,7 +2,7 @@ package com.newrelic.jfr.toevent;
 
 import static com.newrelic.jfr.RecordedObjectValidators.hasField;
 
-import com.newrelic.jfr.profiler.MethodSupport;
+import com.newrelic.jfr.MethodSupport;
 import com.newrelic.telemetry.Attributes;
 import com.newrelic.telemetry.events.Event;
 import java.util.Collections;
