+++ jfr-mappers/src/test/java/com/newrelic/jfr/toevent/ValhallaVBCDetectorTest.java
@@ -12,7 +12,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import com.newrelic.jfr.RecordedObjectValidators;
-import com.newrelic.jfr.profiler.MethodSupport;
+import com.newrelic.jfr.MethodSupport;
 import com.newrelic.telemetry.Attributes;
 import com.newrelic.telemetry.events.Event;
 import java.time.Instant;
